+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/db/repo/StockMoveSupplychainRepository.java
@@ -34,6 +34,9 @@ public class StockMoveSupplychainRepository extends StockMoveManagementRepositor
     if (copy.getStockMoveLineList() != null) {
       for (StockMoveLine stockMoveLine : copy.getStockMoveLineList()) {
         stockMoveLine.setReservedQty(BigDecimal.ZERO);
+        stockMoveLine.setRequestedReservedQty(BigDecimal.ZERO);
+        stockMoveLine.setIsQtyRequested(false);
+        stockMoveLine.setReservationDateTime(null);
         stockMoveLine.setQtyInvoiced(null);
       }
     }