+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/web/StockMoveLineController.java
@@ -17,8 +17,15 @@
  */
 package com.axelor.apps.supplychain.web;
 
+import com.axelor.apps.base.db.Product;
 import com.axelor.apps.stock.db.StockMoveLine;
 import com.axelor.apps.stock.db.repo.StockMoveLineRepository;
+import com.axelor.apps.supplychain.exception.IExceptionMessage;
+import com.axelor.apps.supplychain.service.ReservedQtyService;
+import com.axelor.exception.AxelorException;
+import com.axelor.exception.db.repo.TraceBackRepository;
+import com.axelor.exception.service.TraceBackService;
+import com.axelor.i18n.I18n;
 import com.axelor.inject.Beans;
 import com.axelor.rpc.ActionRequest;
 import com.axelor.rpc.ActionResponse;
@@ -174,5 +181,121 @@ public class StockMoveLineController {
       }
     }
     response.setValue("subLineList", subLines);
+  }
+
+  /**
+   * Called from stock move form view, on clicking allocateAll button on one stock move line. Call
+   * {@link ReservedQtyService#allocateAll(StockMoveLine)}.
+   *
+   * @param request
+   * @param response
+   */
+  public void allocateAll(ActionRequest request, ActionResponse response) {
+    try {
+      StockMoveLine stockMoveLine = request.getContext().asType(StockMoveLine.class);
+      stockMoveLine = Beans.get(StockMoveLineRepository.class).find(stockMoveLine.getId());
+      Product product = stockMoveLine.getProduct();
+      if (product == null || !product.getStockManaged()) {
+        throw new AxelorException(
+            TraceBackRepository.CATEGORY_INCONSISTENCY,
+            I18n.get(IExceptionMessage.SALE_ORDER_LINE_PRODUCT_NOT_STOCK_MANAGED));
+      }
+      Beans.get(ReservedQtyService.class).allocateAll(stockMoveLine);
+      response.setReload(true);
+    } catch (Exception e) {
+      TraceBackService.trace(response, e);
+    }
+  }
+
+  /**
+   * Called from stock move form view, on clicking allocateAll button on one stock move line. Call
+   * {@link ReservedQtyService#updateReservedQty(stockMoveLine, BigDecimal.ZERO)}.
+   *
+   * @param request
+   * @param response
+   */
+  public void deallocateAll(ActionRequest request, ActionResponse response) {
+    try {
+      StockMoveLine stockMoveLine = request.getContext().asType(StockMoveLine.class);
+      stockMoveLine = Beans.get(StockMoveLineRepository.class).find(stockMoveLine.getId());
+      Beans.get(ReservedQtyService.class).updateReservedQty(stockMoveLine, BigDecimal.ZERO);
+      response.setReload(true);
+    } catch (Exception e) {
+      TraceBackService.trace(response, e);
+    }
+  }
+
+  /**
+   * Called from stock move line form view, on request qty click. Call {@link
+   * ReservedQtyService#requestQty(StockMoveLine)}.
+   *
+   * @param request
+   * @param response
+   */
+  public void requestQty(ActionRequest request, ActionResponse response) {
+    try {
+      StockMoveLine stockMoveLine = request.getContext().asType(StockMoveLine.class);
+      stockMoveLine = Beans.get(StockMoveLineRepository.class).find(stockMoveLine.getId());
+      Product product = stockMoveLine.getProduct();
+      if (product == null || !product.getStockManaged()) {
+        throw new AxelorException(
+            TraceBackRepository.CATEGORY_INCONSISTENCY,
+            I18n.get(IExceptionMessage.SALE_ORDER_LINE_PRODUCT_NOT_STOCK_MANAGED));
+      }
+      Beans.get(ReservedQtyService.class).requestQty(stockMoveLine);
+      response.setReload(true);
+    } catch (Exception e) {
+      TraceBackService.trace(response, e);
+    }
+  }
+
+  /**
+   * Called from stock move line form view, on request qty click. Call {@link
+   * ReservedQtyService#cancelReservation(StockMoveLine)}.
+   *
+   * @param request
+   * @param response
+   */
+  public void cancelReservation(ActionRequest request, ActionResponse response) {
+    try {
+      StockMoveLine stockMoveLine = request.getContext().asType(StockMoveLine.class);
+      stockMoveLine = Beans.get(StockMoveLineRepository.class).find(stockMoveLine.getId());
+      Product product = stockMoveLine.getProduct();
+      if (product == null || !product.getStockManaged()) {
+        throw new AxelorException(
+            TraceBackRepository.CATEGORY_INCONSISTENCY,
+            I18n.get(IExceptionMessage.SALE_ORDER_LINE_PRODUCT_NOT_STOCK_MANAGED));
+      }
+      Beans.get(ReservedQtyService.class).cancelReservation(stockMoveLine);
+      response.setReload(true);
+    } catch (Exception e) {
+      TraceBackService.trace(response, e);
+    }
+  }
+
+  /**
+   * Called from stock move line request quantity wizard view. Call {@link
+   * ReservedQtyService#updateReservedQty(StockMoveLine, BigDecimal)}.
+   *
+   * @param request
+   * @param response
+   */
+  public void changeReservedQty(ActionRequest request, ActionResponse response) {
+    try {
+      StockMoveLine stockMoveLine = request.getContext().asType(StockMoveLine.class);
+      BigDecimal newReservedQty = stockMoveLine.getReservedQty();
+      stockMoveLine = Beans.get(StockMoveLineRepository.class).find(stockMoveLine.getId());
+
+      Product product = stockMoveLine.getProduct();
+      if (product == null || !product.getStockManaged()) {
+        throw new AxelorException(
+            TraceBackRepository.CATEGORY_INCONSISTENCY,
+            I18n.get(IExceptionMessage.SALE_ORDER_LINE_PRODUCT_NOT_STOCK_MANAGED));
+      }
+      Beans.get(ReservedQtyService.class).updateReservedQty(stockMoveLine, newReservedQty);
+      response.setReload(true);
+    } catch (Exception e) {
+      TraceBackService.trace(response, e);
+    }
   }
 }